+++ src/components/Sidebar/Sidebar.tsx
@@ -1,7 +1,11 @@
 import { FilePlus, FolderPlus, Home, Share2, Files } from "lucide-react";
+import { useContext } from "react";
 import { Link } from "react-router-dom";
+import { AVAILABLE_DIALOGS, FilesDialogsContext } from "../../context/index";
 
 export function Sidebar() {
+  const { openDialog } = useContext(FilesDialogsContext);
+
   return (
     <aside className="col-span-2 row-span-5 h-full bg-gray-200 p-4">
       <div className="flex flex-col justify-between space-y-8">
@@ -17,6 +21,9 @@ export function Sidebar() {
           <button
             className="flex w-full items-center justify-center gap-1.5 rounded-full bg-blue-600 px-4 py-2 text-white hover:bg-blue-700"
             aria-label="New Folder"
+            onClick={() => {
+              openDialog(AVAILABLE_DIALOGS.CREATE_FOLDER, null);
+            }}
           >
             <FolderPlus className="min-w-6 min-h-6" />
             <span className="hidden md:inline">New Folder</span>
@@ -50,3 +57,11 @@ export function Sidebar() {
     </aside>
   );
 }
+
+/* 
+<CreateFolderDialog
+  isOpen={modalIsOpen}
+  currentParentDirectory={directory}
+  closeModalCallback={closeModal}
+/>
+*/